update register, rename MainConteiner
